third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-01-expected.txt

CONSOLE ERROR: Loading the script 'http://localhost:8000/security/contentSecurityPolicy/resources/script.js' violates the following Content Security Policy directive: "script-src 'self'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback. The action has been blocked.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
